Cleaned up frontend code, switched from monolothic architecture to microservices
