+++ frontend/src/components/escuelas/schoolSampleData.ts
@@ -1,0 +1,93 @@
+import { SchoolStructure } from "./escuelasInterfaces";
+
+export const mockSchools: SchoolStructure[] = [
+  {
+    id: 1,
+    region: "Guadalajara",
+    school: "Primaria",
+    name: "Benito Juárez",
+    employees: "22",
+    students: "410",
+    level: "Primaria",
+    cct: "14EPR1234A",
+    mode: "Escolarizada",
+    shift: "Matutino",
+    address: "Av. Reforma 123",
+    location: "https://maps.google.com/?q=20.674,-103.344",
+    type: "Publica",
+    category: "Infraestructura",
+    notes: "Requiere mejora de sanitarios.",
+    progress: 6500,
+    needs: [
+      {
+        id: 1,
+        itemName: "Pintura para aulas",
+        quantity: 30,
+        unit: "Cubetas",
+        amount: 3000,
+      },
+      {
+        id: 2,
+        itemName: "Reparación de baños",
+        quantity: 1,
+        unit: "Servicio",
+        amount: 7000,
+      },
+    ],
+  },
+  {
+    id: 2,
+    region: "Puerto Vallarta",
+    school: "Secundaria",
+    name: "Lázaro Cárdenas",
+    employees: "31",
+    students: "560",
+    level: "Secundaria",
+    cct: "14ESN9123K",
+    mode: "Escolarizada",
+    shift: "Vespertino",
+    address: "Calle Mar 45",
+    location: "https://maps.google.com/?q=20.653,-105.225",
+    type: "Publica",
+    category: "Materiales Educativos",
+    notes: "Necesita equipamiento para laboratorio.",
+    progress: 7200,
+    needs: [
+      {
+        id: 1,
+        itemName: "Microscopios",
+        quantity: 8,
+        unit: "Piezas",
+        amount: 12000,
+      },
+      { id: 2, itemName: "Reactivos", quantity: 1, unit: "Lote", amount: 3000 },
+    ],
+  },
+  {
+    id: 3,
+    region: "Zapopan",
+    school: "Preescolar",
+    name: "Niños Héroes",
+    employees: "14",
+    students: "180",
+    level: "Preescolar",
+    cct: "14EJN4512P",
+    mode: "Escolarizada",
+    shift: "Matutino",
+    address: "Camino Real 222",
+    location: "https://maps.google.com/?q=20.717,-103.391",
+    type: "Privada",
+    category: "Tecnología",
+    notes: "Proyecto completado.",
+    progress: 12000,
+    needs: [
+      {
+        id: 1,
+        itemName: "Tabletas educativas",
+        quantity: 20,
+        unit: "Piezas",
+        amount: 12000,
+      },
+    ],
+  },
+];
