+++ frontend/src/components/main/MainDashboard.tsx
@@ -1,80 +1,41 @@
-import { useState, useEffect } from 'react';
-import { Box, Typography, Paper, Grid } from '@mui/material';
-import { useRouter } from 'next/navigation';
-import {
-  ReceiptLongOutlined,
-  InventoryOutlined,
-  SecurityOutlined,
-} from '@mui/icons-material';
+"use client";
 
-interface NavigationCard {
-  title: string;
-  description: string;
-  icon: React.ElementType;
-  href: string;
-  color: string;
-  hoverColor: string;
-}
+import { Box, Typography, useTheme, useMediaQuery } from "@mui/material";
+import MainCards from "./StatsCards";
+import RecentSchoolsTable from "./RecentSchools";
+import RecentDonationsTable from "./RecentDonations";
 
-const navigationCards: NavigationCard[] = [
-  {
-    title: 'Ventas',
-    description: 'Gestiona tus ventas, tickets y reportes',
-    icon: ReceiptLongOutlined,
-    href: '/ventas',
-    color: '#0B4595',
-    hoverColor: '#014196',
-  },
-  {
-    title: 'Inventario',
-    description: 'Controla tu stock y productos',
-    icon: InventoryOutlined,
-    href: '/inventario',
-    color: '#0B4595',
-    hoverColor: '#014196',
-  },
-  {
-    title: 'Auditoría',
-    description: 'Revisa el historial de actividades',
-    icon: SecurityOutlined,
-    href: '/auditoria',
-    color: '#0B4595',
-    hoverColor: '#014196',
-  },
-];
-
-export default function MainDashboard() {
-  const router = useRouter();
-  const [userName, setUserName] = useState<string>('');
-  const [greeting, setGreeting] = useState<string>('');
-
-  useEffect(() => {
-    setUserName('Usuario');
-
-    // Set greeting based on time of day
-    const hour = new Date().getHours();
-    if (hour < 12) {
-      setGreeting('Buenos días');
-    } else if (hour < 19) {
-      setGreeting('Buenas tardes');
-    } else {
-      setGreeting('Buenas noches');
-    }
-  }, []);
-
-  const handleCardClick = (href: string) => {
-    router.push(href);
-  };
+export default function DashboardPage() {
+  const theme = useTheme();
+  const isMobile = useMediaQuery(theme.breakpoints.down("sm"));
 
   return (
-    <Box
-      sx={{
-        minHeight: '100vh',
-        backgroundColor: '#F1F5F9',
-        padding: 4,
-      }}
-    >
-        <h1>Bienvenido a ImpulsaEdu</h1>
+    <Box>
+      {/* Welcome Section */}
+      <Box sx={{ marginBottom: 4 }}>
+        <Typography
+          variant="h4"
+          sx={{
+            fontWeight: 700,
+            marginBottom: 1,
+            color: "#1c3661",
+          }}
+        >
+          Bienvenido, Juan
+        </Typography>
+        <Typography sx={{ color: "#4a5f8f" }}>
+          Aquí tienes una descripción general de tus donaciones y escuelas
+        </Typography>
+      </Box>
+
+      {/* Main Stats Cards */}
+      <MainCards />
+
+      {/* Recent Schools Section */}
+      <RecentSchoolsTable />
+
+      {/* Recent Donations Section */}
+      <RecentDonationsTable />
     </Box>
   );
 }
